Switch all funcs to use options struct
Using functional options is nice for simple functions, but
doesn't really work well when we want to compose lots of
options together (like having a func that takes options
related to tags AND options related to images, and wanting
to reuse those for other funcs).

Instead just pass in a (possibly-nil) pointer to a struct
of options. Common config can be reused by just embedding
the structs.

This is a breaking change, but nothing has been tagged
yet, so.......